% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSMOTE.R
\name{ConfusionMatrix}
\alias{ConfusionMatrix}
\title{Confusion Matrix}
\usage{
ConfusionMatrix(pred, target)
}
\arguments{
\item{pred}{the prediction of model.}

\item{target}{the true label.}
}
\value{
the confusion matrix.
}
\description{
Confusion Matrix
}
