% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSMOTE.R
\name{distance}
\alias{distance}
\title{distance in kernel space}
\usage{
distance(x, y)
}
\arguments{
\item{x}{is a vertor in feature space}

\item{y}{is a vertor in feature space}
}
\value{
distance in kernel space
}
\description{
This is a function define distance in kernel space.
}
