% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSMOTE.R
\name{K_SMOTE}
\alias{K_SMOTE}
\title{K_SMOTE}
\usage{
K_SMOTE(train_set, test_set, P, k)
}
\arguments{
\item{train_set}{is the train set.}

\item{test_set}{is the test set.}

\item{P}{is the number of generating samples.}

\item{k}{is the k of kNN algorithm.}
}
\value{
output is a list of two matrix train_K_mat and test_K_mat.
}
\description{
This a function to generate kernel matrix on train_set and test_set.
}
