% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSMOTE.R
\name{kernel_poly}
\alias{kernel_poly}
\title{kernel function}
\usage{
kernel_poly(x, y)
}
\arguments{
\item{x}{is a vector in feature space}

\item{y}{is a vector}
}
\value{
kernel inner product
}
\description{
This is some descriptio of this function.
}
