% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSMOTE.R
\name{modify_label}
\alias{modify_label}
\title{modify_label}
\usage{
modify_label(train_set, test_set, P)
}
\arguments{
\item{train_set}{is the train set with target.}

\item{test_set}{is the test set with target.}

\item{P}{is the number of generating samples.}
}
\value{
output is a list of modified target.
}
\description{
modify_label
}
