% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSMOTE.R
\name{SMO_SVC}
\alias{SMO_SVC}
\title{SMO_SVC}
\usage{
SMO_SVC(
  train_mat,
  test_mat,
  train_target,
  test_target,
  iter_max = 10,
  epsilon = 0.01,
  C = 1
)
}
\arguments{
\item{train_mat}{the kernel matrix of train set.}

\item{test_mat}{the kernel matrix of test set.}

\item{train_target}{train set target.}

\item{test_target}{test set target.}

\item{iter_max}{the maximum of iteration.}

\item{epsilon}{a next condition.}

\item{C}{the parameter of wrong.}
}
\value{
the predict result.
}
\description{
SMO_SVC
}
