\name{StatComp-package}
\alias{StatComp-package}
\docType{package}
\title{
This package provides several functions for benchmarking R and Rcpp.
}
\description{
R and Rcpp versions of gibbs sampler and vaac function are provided in this package.
}
\author{
Yu Ji
Maintainer: Yu Ji <jiyu2017@mail.ustc.edu.cn>
}
\keyword{ package }
