% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSMOTE.R
\name{update_alpha_2}
\alias{update_alpha_2}
\title{update alpha}
\usage{
update_alpha_2(alpha_2, L, H)
}
\arguments{
\item{alpha_2}{old alpha.}

\item{L}{the low.}

\item{H}{the high.}
}
\value{
output is alpha_new.
}
\description{
update alpha
}
